Question is in this file
An Array of numbers  N is given to you. Each element in the array is a single-digit between 0 to 9 only.
Like  N[1 ,2 ,3 ,4 ,5 ]
you have to consider it as 12345 as a number. But it is an array.
what you have to do is to add 1 to the last element of the array.
like input is N[1 ,2 ,3 ,4 ,5 ]
and output is N[1 ,2 ,3 ,4 ,6 ]